Create an assertion function

This function creates an assertion function that can be used to check the validity of an input. All assertions provided with this package are created using either assert_create() or assert_create_advanced()

Usage

assert_create(func, default_error_msg)

Arguments

func

A function defining the assertion criteria. This function should return a logical value (TRUE when assertion is passed or FALSE when it fails)

A

character string providing an error message in case the assertion fails. Can include the following special terms

{arg_name} to refer to the name of the variable supplied to the assertion.

{arg_value} to refer to the value of the variable supplied to the assertion

{code_to_evaluate} to evaluate the code within the error message. Replace code_to_evaluate with your code

{.strong bold_text} to perform inline formatting. Replace bold_text with your text. See cli documentation for details

Value

An assertion function.

Examples

if (FALSE) {
# Create an assertion function that checks that a character string is all
# lower case
assert_character <- assert_create(is.character, "{arg} must be a character vector, not a {class(x)}")

# Use the assertion function
is_lower("hello") # Returns invisible TRUE
is_lower("Hello") # Aborts the function with the error message
}